+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/raulsebastianmihaila/create-crizmas/compare/v1.2.0...v2.0.0) (2021-01-01)
+
+### Breaking changes
+- Update crizmas related packages to v2.
+
+### Fixes
+- Fix importing the css file in all the versions.
+
+### Updates
+- Enable tree shaking with webpack for crizmas related packages.
+
 <a name="1.2.0"></a>
 # [1.2.0](https://github.com/raulsebastianmihaila/create-crizmas/compare/v1.1.0...v1.2.0) (2020-11-30)